requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
Pillow==6.2.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2019.3
six==1.13.0
style==1.1.0
update==0.0.1
whitenoise==4.1.2
